requirements.txt
transformers==4.46.2
autoawq
auto-gptq
bitsandbytes
sacrebleu
unbabel-comet